Works/Homework1/requirements.txt
beautifulsoup4
requests
pyzmq

